requirements.txt
asgiref==3.7.2
attrs==23.1.0
decouple==0.0.7
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rest-framework-simplejwt==0.0.2
rpds-py==0.9.2
sqlparse==0.4.4
uritemplate==4.1.1